+++ front/src/index.jsx
@@ -1,5 +1,8 @@
 import React from 'react';
 import {render} from 'react-dom';
+import css from './style.scss';
+
+console.log(css);
 
 class App extends React.Component {
 	constructor(props) {
@@ -13,12 +16,14 @@ class App extends React.Component {
 	render() {
 		return (
 			<div>
-				<input type="text" onChange = { this.onChange.bind(this)} />
+				<input type="text" className="input is-info" placeholder="Please input text" onChange = { this.onChange.bind(this)} />
 				<p>{ this.state.message }</p>
 			</div>
 		)
 	}
 }
+
+//require('./style.scss');
 
 render(<App/>, document.getElementById('app'));
