+++ front/src/style.scss
@@ -1,0 +1,2 @@
+@charset "utf-8";
+@import "../node_modules/bulma/bulma.sass";
